requirements.txt
fastmcp
psutil
pyperclip
pyautogui
pygetwindow
opencv-python
Pillow
plyer
schedule
SpeechRecognition
pyttsx3
pyaudio
